Update README.md
Removed the documentation section. As the doc is generated from this git repo, this section is redundant. Also when you click on the "documentation site" link in this line "Please refer to the AWS CDK Observability Accelerator documentation site for complete project documentation.", it looks back to the same page.